Refact: Move "Auth", "SearchTarget" Model Package
<description>
- Auth, SearchTarget 데이터 모델 패키지 변경

<issue>
NGOL-92
