+++ common/data/src/main/java/com/nugu/nuguollim/common/data/model/auth/SignUpData.kt
@@ -1,0 +1,8 @@
+package com.nugu.nuguollim.common.data.model.auth
+
+data class SignUpData(
+    val id: Int,
+    val providerType: String,
+    val providerId: String,
+    val nickname: String,
+)
