+++ common/data/src/main/java/com/nugu/nuguollim/common/data/model/auth/TokenData.kt
@@ -1,0 +1,3 @@
+package com.nugu.nuguollim.common.data.model.auth
+
+data class TokenData(val token: String)
