--- core/data/src/main/java/com/nuguollim/data/model/auth/TokenData.kt
@@ -1,9 +1,0 @@
-package com.nuguollim.data.model.auth
-
-import com.nuguollim.remote.model.auth.TokenResponse
-
-data class TokenData(val token: String) {
-    companion object {
-        fun TokenResponse.asExternalModel(): TokenData = TokenData(accessToken)
-    }
-}
