+++ core/data/src/main/java/com/nuguollim/data/repository/auth/AuthRepositoryImpl.kt
@@ -1,12 +1,12 @@
 package com.nuguollim.data.repository.auth
 
 import com.nugu.data_store.data_source.auth.AuthLocalDataSource
-import com.nuguollim.data.model.auth.AuthInfo
-import com.nuguollim.data.model.auth.SignUpData
-import com.nuguollim.data.model.auth.SignUpData.Companion.asExternalModel
-import com.nuguollim.data.model.auth.TokenData
-import com.nuguollim.data.model.auth.TokenData.Companion.asExternalModel
+import com.nugu.nuguollim.common.data.model.auth.AuthInfo
+import com.nugu.nuguollim.common.data.model.auth.SignUpData
+import com.nugu.nuguollim.common.data.model.auth.TokenData
 import com.nuguollim.remote.data_source.auth.AuthRemoteDataSource
+import com.nuguollim.remote.model.auth.SignupResponse.Companion.asExternalModel
+import com.nuguollim.remote.model.auth.TokenResponse.Companion.asExternalModel
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import okhttp3.RequestBody
