+++ core/data/src/main/java/com/nuguollim/data/repository/seach/SearchRepository.kt
@@ -1,6 +1,6 @@
 package com.nuguollim.data.repository.seach
 
-import com.nuguollim.data.model.search.target.TemplateTargetData
+import com.nugu.nuguollim.common.data.model.search.target.TemplateTargetData
 import kotlinx.coroutines.flow.Flow
 
 interface SearchRepository {
