+++ core/data/src/main/java/com/nuguollim/data/repository/seach/SearchRepositoryImpl.kt
@@ -1,8 +1,8 @@
 package com.nuguollim.data.repository.seach
 
-import com.nuguollim.data.model.search.target.TemplateTargetData
-import com.nuguollim.data.model.search.target.TemplateTargetData.Companion.asExternalModel
+import com.nugu.nuguollim.common.data.model.search.target.TemplateTargetData
 import com.nuguollim.remote.data_source.search.SearchRemoteDataSource
+import com.nuguollim.remote.model.search.target.TemplateTargetResponse.Companion.asExternalModel
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
