+++ core/data/src/main/java/com/nuguollim/data/usecase/auth/GetAuthInfoUseCase.kt
@@ -1,6 +1,6 @@
 package com.nuguollim.data.usecase.auth
 
-import com.nuguollim.data.model.auth.AuthInfo
+import com.nugu.nuguollim.common.data.model.auth.AuthInfo
 import com.nuguollim.data.repository.auth.AuthRepository
 import com.nuguollim.data.state.ResultState
 import com.nuguollim.data.state.resultStateFlow
