+++ core/remote/src/main/java/com/nuguollim/remote/model/auth/SignupResponse.kt
@@ -1,8 +1,19 @@
 package com.nuguollim.remote.model.auth
 
+import com.nugu.nuguollim.common.data.model.auth.SignUpData
+
 data class SignupResponse(
     val id: Int,
     val providerType: String,
     val providerId: String,
     val nickname: String,
-)
+) {
+    companion object {
+        fun SignupResponse.asExternalModel(): SignUpData = SignUpData(
+            id = id,
+            providerType = providerType,
+            providerId = providerId,
+            nickname = nickname,
+        )
+    }
+}
