+++ core/remote/src/main/java/com/nuguollim/remote/model/auth/TokenResponse.kt
@@ -1,3 +1,9 @@
 package com.nuguollim.remote.model.auth
 
-data class TokenResponse(val accessToken: String)
+import com.nugu.nuguollim.common.data.model.auth.TokenData
+
+data class TokenResponse(val accessToken: String) {
+    companion object {
+        fun TokenResponse.asExternalModel(): TokenData = TokenData(accessToken)
+    }
+}
